Add HTTP integration tests, lib target, and README
- Split the backend into a lib + thin binary so integration tests can
  drive the real Axum router
- tests/api.rs: onboarding + username uniqueness, a full two-player game
  over HTTP (opening solo move, challenge, turn handoff, invalid-word 422,
  not-your-turn, resign), invite preview/accept/claim, and shape-proxy
  authorization (unknown view, non-participant 403, enforced where-clause,
  game_secrets never reachable). Auth uses locally minted RS256 JWTs.
- Fix: length() returns INT4, decode rack length as i32
- README with architecture overview, local dev, and infra setup checklist

+++ backend/src/handlers/moves.rs
@@ -148,7 +148,7 @@ pub async fn make_move(
     } else {
         let next_player = opponent_id.clone().expect("active game has opponent");
         let next_rack_empty: bool = {
-            let n: Option<i64> = sqlx::query_scalar(
+            let n: Option<i32> = sqlx::query_scalar(
                 "SELECT length(rack) FROM game_players WHERE game_id = $1 AND user_id = $2",
             )
             .bind(id)
